chore: move the one-decimal ratio to its own pull request
This reverts commit dcd2c55.

+++ api/src/dto/alliance/war/dto_statistic.py
@@ -20,7 +20,7 @@ class PlayerSeasonStatsResponse(BaseModel):
     total_miniboss: int
     total_boss: int
     total_not_fought: int
-    ratio: float
+    ratio: int
     wars_participated: int
     avg_fights_per_war: float
     avg_boss_miniboss_per_war: float
